pyproject.toml
[tool.poetry]
name = "airbyte-api-sdk"
version = "0.1.0"
description = ""
authors = ["bhavaniravi <bhava0895@gmail.com>"]
readme = "README.md"
packages = [{include = "airbyte_api_sdk"}]

[tool.poetry.dependencies]
python = "^3.11"
airbyte-api = "^0.6.0"
requests-toolbelt = "^1.0.0"
python-dotenv = "^1.0.0"
black = "22.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
